LayoutTests/http/wpt/wk-web-animations/interfaces/keyframe-effect-getComputedStyle.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Using getComputedStyle() during animations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<div id="hw-target"></div>
<div id="sw-target"></div>
<script>
'use strict';

testRunner.waitUntilDone();

const tolerance = 1/100;

const hwTest = async_test("Querying the computed style for an accelerated animation.");
const swTest = async_test("Querying the computed style for a non-accelerated animation.");

const hwAnimation = new Animation(new KeyframeEffect(document.getElementById("hw-target"), [
    { opacity: 0 },
    { opacity: 1 }
]));
hwAnimation.startTime = 0;
hwAnimation.effect.timing.duration = 10000;

const swAnimation = new Animation(new KeyframeEffect(document.getElementById("sw-target"), [
    { left: 0 },
    { left: "100px" }
]));
swAnimation.startTime = 0;
swAnimation.effect.timing.duration = 10000;

function progress(animation) {
    return (animation.currentTime - animation.startTime) / animation.effect.timing.duration;
}

hwTest.step(() => {
    assert_greater_than(hwAnimation.currentTime, hwAnimation.startTime, "The animation has started.");
    assert_approx_equals(Number(getComputedStyle(hwAnimation.effect.target).opacity), progress(hwAnimation), tolerance);
});
hwTest.done();

swTest.step(() => {
    assert_greater_than(swAnimation.currentTime, swAnimation.startTime, "The animation has started.");
    assert_approx_equals(parseFloat(getComputedStyle(swAnimation.effect.target).left), progress(swAnimation) * 100, tolerance);
});
swTest.done();

testRunner.notifyDone();

</script>
</body>
